Create Core Library.
The core library is what firewall backends will use to inherit
and get powers to control the system. The inheritance should
use IFirewallClientBackend, that needs to be registered via a
FirewallClient, which in turn is what's going to be used in
Qml.

All calls to FirewallClient will just be forwarded to the internal
IFirewallClientBackend, with the exception of signals: the inner
class should trigger the parent's signals as it's what's connected
in the Qml.

Log represents a line of log from the firewall
Rule represents a firewal rule (incoming or outcoming)

Most of those are abstracted enough for backend implementations.