Remove KAuth based code
This is a bit bigger than I wanted but I did not see how to easily
break this into smaller steps without redoing most of the work.
This remoes all calls to KAuth from netstat, and also removes the
`netstat` binary call - it's now only supposed to work with SS
Perhaps I should change the name of the files.