Fix Logs visualization for UFW
This was intense. At first the logs where not showing at all
because UFW does not use a log file by default, but it uses
systemd. I think I can safely assume that I want to use systemd
to get the information from journalctl.

Then I realied that the regexp was wrong, and not getting
any information - usually I dislike regexps because they
tend to be not really clear and just a gibberish, so I ported
to a sligth bigger, but easier to understand, code.

Third, the amount of log was too much - 12.000 lines?
This is not a professional tool, for those using the
logs from the system is better, I'v set a cap of 100 log lines
- we can later on use a preference system for that.