KCM: Port Kirigami.OverlaySheet to new API, refactor both sheets
About dialog now will be created on demand, improving start-up time.
Its properties are set via function instead of bindings to mitigate
non-NOTIFYable property of kcm.client object. It's layout does not
look like what it is intended to be, because having a FormLayout with
SelectableLabel in it requires some fixing on Kirigami side, both in
FormLayout and in OverlaySheet. Hopefully it will all be sorted out by
the time Plasma 6.0 comes out; but meanwhile it's not the end of the
world either.

In the main component, "drawer" OverlaySheet for adding and editing
existing rules has had some changes too. Its event handling for
Enter/Return and Escape keys was simplified, thanks to porting of
OverlaySheet to T.Popup which fixed tons of focus-related bugs on its
own. Its DialogButtonBox with the only button was moved to the footer
of OverlaySheet for better UI and semantic integration.

This patch does not make the KCM usable as a whole, but you can launch
it if you comment out QQC1 import and all the QQC1.TableView component.