netstat: simplify timeout handling
this was a bit convoluted. in practice this simply amounted to the
helper terminating ss after 10 seconds - a feat easily achieved by
stopping on the next query if it doesn't finish in time. no need for
kill timers, the ConnectionsModel already has a timer.

Fixes SYSTEMSETTINGS-5H
(by virtue of not having the code anymore)