kcm: hide disabled actions instead of disabling them
Normally we do want to just disable actions, but in this case they're
really rather distracting when disabled, and can confuse the eye which
is looking for the switch to enable the firewall. I noticed the
Bluetooth KCM does the same thing. So let's hide those actions here as
well.