Mark KCM as needing changes after modifications
Firewalld has two backends, UFW and
firewalld. UFW is instant apply, firewalld is complex.

It immediately makes polkit calls, but then has a final step to
tell firewalld to commit changes. This last step requires KCM::save
to be called, which requires the KCM to be marked as needing changes if
any previous calls to firewalld were made.